Ajout de Telecom Investigation (ability) Suppression de l’ancienne. Ajout de l’ability as view dans le init.py (abilities)
