+++ pychronia_game/views/abilities/__init__.py
@@ -21,10 +21,10 @@ matter_analysis = MatterAnalysisAbility.as_view
 from .world_scan_mod import WorldScanAbility
 world_scan = WorldScanAbility.as_view
 
-'''
+
 from .telecom_investigation_mod import TelecomInvestigationAbility
 telecom_investigation = TelecomInvestigationAbility.as_view
-'''
+
 
 from .artificial_intelligence_mod import ArtificialIntelligenceAbility
 artificial_intelligence = ArtificialIntelligenceAbility.as_view
