+++ pychronia_game/views/abilities/telecom_investigation_mod.py
@@ -1,0 +1,56 @@
+# -*- coding: utf-8 -*-
+from __future__ import print_function
+from __future__ import unicode_literals
+
+from pychronia_game.common import *
+from pychronia_game.datamanager.abstract_ability import AbstractAbility
+from pychronia_game.datamanager.abstract_game_view import register_view
+from pychronia_game.datamanager.datamanager_tools import readonly_method, \
+    transaction_watcher
+from pychronia_game.forms import OtherCharactersForm
+
+@register_view
+class TelecomInvestigationAbility(AbstractAbility):
+
+    TITLE = ugettext_lazy("Telecom Investigation")
+    NAME = "telecom_investigation"
+
+    GAME_ACTIONS = dict(investigation_form=dict(title=ugettext_lazy("Process Telecom Investigation"),
+                                               form_class=OtherCharactersForm,
+                                               callback="process_telecom_investigation"))
+    
+    TEMPLATE = "abilities/telecom_investigation.html"
+
+    ACCESS = UserAccess.authenticated
+    REQUIRES_CHARACTER_PERMISSION = False
+    REQUIRES_GLOBAL_PERMISSION = False
+
+    @classmethod
+    def _setup_ability_settings(cls, settings):
+        pass # nothing to do
+
+    def _setup_private_ability_data(self, private_data):
+        pass # nothing to do
+
+    def _check_data_sanity(self, strict=False):
+        pass # nothing to do
+
+    def get_template_vars(self, previous_form_data=None):
+        translation_form = self._instantiate_game_form(new_action_name="investigation_form",
+                                                       hide_on_success=False,
+                                                      previous_form_data=previous_form_data)
+        translation_delay = (2,3)
+        return {
+                 'page_title': _("Telecom Investigation"),
+                 "investigation_form": translation_form,
+                 'min_delay_mn': translation_delay[0],
+                 'max_delay_mn': translation_delay[1],
+               }
+
+
+    
+    @staticmethod
+    def process_telecom_investigation(self):
+        return _("Telecom is in process, you will receive an e-mail with the intercepted messages soon!")
+
+
